added createbusiness

+++ controllers/api/reviewRoutes.js
@@ -4,19 +4,19 @@ const sequelize = require('sequelize')
 
 const withAuth = require('../../utils/auth');
 
-router.post('/:id',  async (req, res) => {
+router.post('/:id', withAuth,  async (req, res) => {
     try {
-        // if (!req.session.loggedIn) {
-        //     res.redirect('/login')
-        //     return
-        // }
+        if (!req.session.loggedIn) {
+            res.redirect('/login')
+            return
+        }
         const newReviewData = await Reviews.create({
-            review: req.body.review,
-            rating: req.body.rating,
+            review: req.body.comment,
+            rating: req.body.starRating,
             user_id: req.session.user_id,
             business_id: req.params.id,
         });
-        res.status(200).redirect(`/business/${req.params.id}`);
+        res.status(200).json();
     } catch (error) {
         console.log(error);
         res.status(500).json(error);
@@ -24,12 +24,12 @@ router.post('/:id',  async (req, res) => {
 });
 
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', withAuth, async (req, res) => {
     try {
-        // if (!req.session.loggedIn) {
-        //     res.redirect('/login')
-        //     return
-        // }
+        if (!req.session.loggedIn) {
+            res.redirect('/login')
+            return
+        }
         const deletedReview = await Reviews.destroy({
             where: {
                 review_id: req.params.id,
